+++ web/src/context/ConfigContext.tsx
@@ -1,4 +1,4 @@
 import { createContext } from "react";
-import { Config } from "../providers/ConfigProvider.tsx";
+import { Config } from "../providers/ConfigProvider";
 
 export const ConfigContext = createContext<Config>({} as unknown as Config);