+++ web/src/hooks/useConfig.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
-import { ConfigContext } from "../context/ConfigContext.tsx";
-import { Config } from "../providers/ConfigProvider.tsx";
+import { ConfigContext } from "../context/ConfigContext";
+import { Config } from "../providers/ConfigProvider";
 
 export function useConfig(): Config {
   return useContext(ConfigContext);