+++ web/src/hooks/useStreamStart.tsx
@@ -1,5 +1,4 @@
 import { notifications } from "@mantine/notifications";
-import { IconExclamationMark } from "@tabler/icons-react";
 import { useMutation } from "@tanstack/react-query";
 import { startSession } from "../state/Sessions";
 import { useConfig } from "./useConfig";
@@ -10,19 +9,29 @@ export default function useStreamStart(appId: string) {
     mutationFn: async () => {
       return await startSession({ config, appId });
     },
+    retry: (failureCount) => {
+      showStreamWarning();
+      return failureCount < 3;
+    },
     onSuccess: (session) => {
       window.location.href = `/app/${appId}/sessions/${session.id}`;
     },
     onError: (error) => {
-      notifications.update({
-        id: streamStartNotification,
-        color: "red",
-        title: "Failed to start a new session",
-        icon: <IconExclamationMark />,
-        loading: false,
-        message: error.message,
-      });
+      console.error("Failed to start a stream:", error);
+      notifications.hide(streamStartNotification);
     },
+  });
+}
+
+export function showStreamWarning() {
+  notifications.hide(streamStartNotification);
+  notifications.show({
+    id: streamStartNotification,
+    loading: true,
+    title: "",
+    message:
+      "Connecting to a streaming session is taking longer than expected, please wait...",
+    autoClose: 30000,
   });
 }